requirements.txt
python-dotenv>=0.19.0
torch
transformers
requests
streamlit
Pillow
pytorch-lightning==2.3.2
lightning-utilities==0.10.1
neuralforecast==3.0.2
utilsforecast==0.2.12
coreforecast==0.0.13